feature-A001-code-refactoring Introduced testing dependencies, improved test coverage
I added some positive and negative test cases to document the current behavior of the system and provide a baseline for sanity checks.
Some edge cases appear to not be handled properly, but it will be addressed later.

P.S. This is also the point where I had to patch and undo A LOT of my progress just so I can go back in commit history and amend the "Simple Maven transformation" commit. I didn't notice most of my files not getting picked up by the git add when I moved them to new packages, so I had to go back and fix it. Oooof >_<