feature-A001-code-refactoring Split rental logic into computing and formatting parts
- I've split the code into Service and Formatter, communicating through Interfaces;
- Utilized Lombok to enhance the data classes;
- Changed POM to be a child of the spring boot starter pom, configured it to work with Lombok and Mockito;
- Rewrote tests to target the now separate compute and format flows;
- Rewrote execution logic to follow a more functional approach with smaller higher cohesion methods;
- Added rudimentary multithreading and ordering to the movie record processing loop;
- Removed the outdated Movie data class and conversion to it, since it doesn't seem to be required and the program is growing complex.

P.S. This commit is on the larger side. In hindsight, I should have done at least one commit in-between tagged as WIP since back then I only had the tests and not the code, but the amount of changes was already large... Also, Git continues to struggle keeping track of files when I move them around, but I don't want to spend time fixing it in this instance. I hope that is alright!