feature-A001-code-refactoring Addressing further PR feedback (simulated)
- Removed unnecessary Consumer and Supplier interfaces and formated the code;
- Rewrote an unnecessary lambda call;
- Added initialValue for AtomicInteger;
- Demonstrated the use of var (specifically Lombok val) instead of explicit types.

P.S. I am not sure if you use vars on your project, considering that some teams deal with an old monolithic application and may not be used to this functional-ish style, but it has its uses when one wants to streamline the code! And Lombok's "val" helps cut down on writing "final var".