feature-A001-code-refactoring - performed the refactoring assignment
* feature-A001-code-refactoring Refactoring Plan

Modified .gitignore and added a high-level plan with assignment comments to the README.md

* feature-A001-code-refactoring Converted to a simple Maven project

I have the time, so I decided to make it a bit more serious and actually use some of the extra dependencies.

* feature-A001-code-refactoring Introduced testing dependencies, improved test coverage

I added some positive and negative test cases to document the current behavior of the system and provide a baseline for sanity checks.
Some edge cases appear to not be handled properly, but it will be addressed later.

P.S. This is also the point where I had to patch and undo A LOT of my progress just so I can go back in commit history and amend the "Simple Maven transformation" commit. I didn't notice most of my files not getting picked up by the git add when I moved them to new packages, so I had to go back and fix it. Oooof >_<

* feature-A001-code-refactoring Converted to a simple Spring Boot application

- Introduced Spring Boot starter dependencies, which allowed to remove my manual imports.
- Changed file names for the Main and RentalInfo classes to better represent their true purpose.
- Restructured the project structure into separate packages.
- Switched to dependency injection for the Application class.

P.S. Apologies for slightly messy commit history, this is recovered from Patch files due to the aforementioned hiccup with missing files in earlier commit. Should be smooth sailing from now on!

* feature-A001-code-refactoring Converted DTOs to records, used Enum for Movies

* feature-A001-code-refactoring Split rental logic into computing and formatting parts

- I've split the code into Service and Formatter, communicating through Interfaces;
- Utilized Lombok to enhance the data classes;
- Changed POM to be a child of the spring boot starter pom, configured it to work with Lombok and Mockito;
- Rewrote tests to target the now separate compute and format flows;
- Rewrote execution logic to follow a more functional approach with smaller higher cohesion methods;
- Added rudimentary multithreading and ordering to the movie record processing loop;
- Removed the outdated Movie data class and conversion to it, since it doesn't seem to be required and the program is growing complex.

P.S. This commit is on the larger side. In hindsight, I should have done at least one commit in-between tagged as WIP since back then I only had the tests and not the code, but the amount of changes was already large... Also, Git continues to struggle keeping track of files when I move them around, but I don't want to spend time fixing it in this instance. I hope that is alright!

* feature-A001-code-refactoring Moved static logic into RentalInfoHelper

- Static methods were moved to a separate helper class
- Helper class covered with tests
- Renamed some of the variables and methods to promote uniform and explicit naming conventions

* feature-A001-code-refactoring Added Supplier and Consumer to the Main app class

Refactored the code to use a generic supplier and consumer instead of hardcoding the "default" app run logic into the main() method.

* feature-A001-code-refactoring Updated README and created TASK_NOTES

* feature-A001-code-refactoring Containerized the application

Also moved one of the tests into correct folder, cleaned-up TODOs, and updated notes.

* feature-A001-code-refactoring Final refactoring steps

- Separated validation logic into separate Validators that return ValidationResult;
- Created a CustomerUtils class;
- Refactored the main generateStatementForCustomer method to be easier to understand at a glance;
- Switched processing to never use shared state, now communication is via the new CustomerRentalSummary record.

P.S. This is just about everything I wanted to do, so I will be making the PR now.

* feature-A001-code-refactoring Addressing PR feedback (simulated)

- Improved TASK_NOTES.md;
- Covered Utils classes with JavaDocs;
- Considered implementing a Validator Chain for MovieRecord validation.

* feature-A001-code-refactoring Addressing further PR feedback (simulated)

- Removed unnecessary Consumer and Supplier interfaces and formated the code;
- Rewrote an unnecessary lambda call;
- Added initialValue for AtomicInteger;
- Demonstrated the use of var (specifically Lombok val) instead of explicit types.

P.S. I am not sure if you use vars on your project, considering that some teams deal with an old monolithic application and may not be used to this functional-ish style, but it has its uses when one wants to streamline the code! And Lombok's "val" helps cut down on writing "final var".

* feature-A001-code-refactoring Addressing the actual PR feedback

- Rewrote the main loop to utilize the reduce() operation instead of relying on atomic variables;
- Got rid of the MovieEnum that was used to cheaply emulate a Movie repository and instead created a MockMovieRepository with all required data types;
- Updated the tests to reflect these changes as well as provided a few more comments to explicitly elaborate my reasoning.